:truck: rename default layout
